+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/IntelliJPlatformType.kt
@@ -73,15 +73,15 @@ enum class IntelliJPlatformType(
         maven = Coordinates("com.jetbrains.intellij.idea", "ideaIU"),
         installer = Coordinates("idea", "ideaIU"),
     ),
+    IntellijIdea(
+        code = "IU",
+        maven = Coordinates("com.jetbrains.intellij.idea", "idea"),
+        installer = Coordinates("idea", "idea"),
+    ),
     JetBrainsClient(
         code = "JBC",
         maven = null,
         installer = Coordinates("idea/code-with-me", "JetBrainsClient"),
-
-        // https://download.jetbrains.com/idea/ideaIC-2025.1.1.1.tar.gz
-        // https://download.jetbrains.com/idea/code-with-me/JetBrainsClient-251.25410.129.tar.gz
-
-        // https://download.jetbrains.com/idea/code-with-me/JetBrainsClient-251.25410.129.aarch64.dmg
     ),
     MPS(
         code = "MPS",